Ensure that __isChanging gets set to false after errors
This commit addresses #136

Wrapping the call to `super._flushProperties()` in a try{}finally{}
ensures that this.__isChanging gets set back to false even if an
error occurs within `super._flushProperties()`, which will avoid
distracting from the real error by logging reentrancy warnings.

I haven't been able to figure out how to write a good test for this.
If I throw an exception in the render function it fails the test
even thoug that's an expected behavior. As the render process
happens asynchronously, I can't trivially catch the error to avoid
failing the test.