Changes property options to support `converter` (#369)
* Changes property options to support `converter`

Fixes #264

Changes `type` to be only a hint to the `converter` option which has the previous `type` functionality, an object with `toAttribute` and `fromAttribute` or just a function which is `fromAttribute`. In addition to the `value` these functions now also get the property's `type`.

Also provides a default converter that supports `Boolean`, `String`, `Number`, `Object`, and `Array` out of the box.

In addition, numbers and strings now become `null` if their reflected attribute is removed.

* Format

* Address review feedback

* Update CHANGELOG.md

* Clarify documentation about default converter

* Format

* Fix test to use lowercase attribute names

This is working around https://github.com/webcomponents/custom-elements/issues/167.

* Fix test on IE

* Address review feedback

Remove superfluous null checks

* Makes reflection more consistent

Previously, when an attribute changed as a result of a reflecting property changing, the property was prevented from mutating again as can happen when a custom `converter` is used.

Now, the oppose is also true. When a property changes as a result of an attribute changing, the attribute is prevented from mutating again

This change helps ensure that when a user calls `setAttribute`, a `converter.toAttribute` does not cause the attribute to immediately mutate. This is unexpected behavior and this change discourages it.

* Format.

* Address review feedback.

* Address review feedback

Ensure Object/Array properties respect `undefined` (no change to attribute) and `null` (remove attribute) values.