Add @ExportDecoratedItems annotations to query and queryAll decorators.
This tells https://github.com/angular/tsickle to add Closure JS Compiler
@export annotations to properties annotated with these decorators. We
already did this for property and other decorators, but we also need it
for these (due to the way that TypeScript generates decorator code which
involves a static string of the property name).

Also makes LitElement.renderRoot public readonly instead of protected
because the decorators need to access that property. Presumably this
problem did not come up before because tsickle was not producing as
correct types for Closure JS Compiler as it is now (maybe it doesn't
understand the additional indirection we had before).