Add file extensions fo imports where missing. (#595)
I was experimenting with [import maps](https://github.com/WICG/import-maps) in Chrome Canary. It won't be really useful for us until https://crbug.com/928149 (and, practically speaking) https://crbug.com/928149 are implemented, but it's at a really promising state!

This change ensures that when we're importing a file within another package, we give the full filename.