Remove unnecessary type assertions (#599)
* Remove unnecessary type assertions

This is the result of turning on the no-unnecessary-type-assertion tslint pass and running `--fix`. There was one place where I disabled this pass where we needed `!` to note that a variable was definitely  assigned to before it was read.

I like this lint pass because it makes the code a bit easier to read and understand. I will admit though that I'm proposing this change because we use this lint pass downstream, and it's annoying seeing the lint warnings.