Add ugly type assertions for `this` (#659)
* Add ugly type assertions for `this`

Because of https://github.com/google/closure-compiler/issues/3340 – `this` has an unknown type to closure compiler in static methods. In order to be compatible with the Closure Compiler + JSConformance `BanUnknownThis` check, we must explicitly add type annotations wherever we access `this` in a static method.

Yes, this is pretty awful from a legibility perspective. We could improve legibility at the cost of additional size by doing:

```typescript
const self = this as typeof LitElement;
```

I went with the zero-size-cost option here.

This required disabling the unnecessary type assertions lint pass :(