Replaces cssLiteral (removed) with unsafeCSS  (#600)
* Replaced cssLiteral with unsafeCSS in style guide

* Updated unsafeCSS justifcation

* Update docs/_guide/styles.md

Co-Authored-By: hyperpress <hyperpress@users.noreply.github.com>

* Corrected import path to include LitElement

Corrected import path to include LitElement since example imports it along with css.

* Clarify when unsafeCSS should be applied

If I understand the issue, we still need to apply unsafeCSS under certain conditions. I've added safe and unsafe examples hoping this will clarify usage. But there may be a remedy in the works I'm not aware of. Tested both of these and they currently pass.

* Change wording of unsafeCSS stuff

* More edits to unsafe CSS stuff