Delegate updateComplete getter to getUpdateComplete method for TS ES5.
When the target language is ES5, TypeScript does not support calling a
superclass getter, which means it is not possible (or at least very
awkward) to override the `updateComplete` getter in an element (because
a call like `super.updateComplete.then(...)` will not compile). See
https://github.com/microsoft/TypeScript/issues/338.

We now delegate the `updateComplete` getter to a new `getUpdateComplete`
method, which can be safely overriden by users. A little awkward, but
seemed like the least bad solution we could think of.