Make eventOptions decorator compatible with tsickle/Closure compiler.
The @ExportDecoratedItems annotation is required to instruct
https://github.com/angular/tsickle to add an @export annotation for
Closure.

The @export annotation is required because TypeScript's polyfilled
decorator output is not compatible with Closure property renaming
optimizations (since it passes the property name as a string).

Also, decorators must be declared as regular functions, not arrow
functions, because the part of tsickle that recognizes decorators does
not currently expect them to be defined as arrow functions.