Adds `attributeShouldChangeProperty(name, oldValue, value)` protected method.
This method allows for customizing the dirty check applied when attribute values change. The default is to avoid setting a corresponding property when an attribute changes and the old and new values are exactly the same. This is being done to address #699. It's done as an opt-in to avoid a breaking change.