Fixes #965
It is now possible to set a computed reflecting property in another property's setter as long as `requestUpdate` is called before setting the computed property.

Note, this does require a slightly user code modification to address #965 but it addresses the use case without degrading performance by adding per-property tracking (which is the other alternative).