Adds `static get eventProperties` and `@eventProperty`
This provides a way of defining properties that accept event listener functions similar to the platform's `.onclick` property. Element authors should supply an `EventPropertyDeclaration` which requires the `name` to be set to the name of the event. The user can then set `element.onEventProperty = (event) => //handler` and when the element fires a corresponding event, the handler is callled. Under the hood, this property simply calls `addEventListener` for the given event name. Only a single event listener can be added in this way. If the property is re-assigned, the previous listener is removed. This feature is useful to help document what events the element fires, and it also provides a seamless integration point with frameworks (like React) which rely on event properties for hooking up events.